tree-wide: post-restructuring cleanup
